+++ CHANGELOG.md
@@ -322,3 +322,4 @@ Rewritten every thing as a module with `nmanga` namespace.
 **Changes**
 - Make `magick convert` to be `magick` instead since it's deprecated in ImageMagick 7+
 - Properly say how many images is copied without autolevel
+- Force use UTF-8 when opening archive metadata